+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 -   Gutter icon is smaller
+-   Reuse the same decorations object when rendering for better perf
 
 ## 1.24.0 - 2024-10-10